fix migration

+++ migrations/000001_create_metrics_table.up.sql
@@ -10,7 +10,7 @@ create table IF NOT EXISTS metrics
         id SERIAL PRIMARY KEY,
         metric_id varchar(50) NOT NULL UNIQUE,
         type metrics_type NOT NULL,
-        delta integer,
+        delta bigint,
         value double precision,
         created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
         updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
